+++ dozer-sink-clickhouse/src/ddl.rs
@@ -1,6 +1,7 @@
-use dozer_types::log::warn;
 use dozer_types::models::sink::ClickhouseSinkTableOptions;
-use dozer_types::types::{FieldDefinition, FieldType};
+use dozer_types::types::FieldDefinition;
+
+use crate::schema::map_field_to_type;
 
 const DEFAULT_TABLE_ENGINE: &str = "MergeTree()";
 
@@ -69,35 +70,4 @@ pub fn get_create_table_query(
             {sample_by}
             ",
     )
-}
-
-pub fn map_field_to_type(field: &FieldDefinition) -> String {
-    let typ = match field.typ {
-        FieldType::UInt => "UInt64",
-        FieldType::U128 => "UInt128",
-        FieldType::Int => "Int64",
-        FieldType::I128 => "Int128",
-        FieldType::Float => "Float64",
-        FieldType::Boolean => "Boolean",
-        FieldType::String => "String",
-        FieldType::Text => "String",
-        FieldType::Binary => "Array(UInt8)",
-        FieldType::Decimal => "Decimal64",
-        FieldType::Timestamp => "DateTime64(3)",
-        FieldType::Date => "Date",
-        FieldType::Json => "JSON",
-        FieldType::Point => "Point",
-        FieldType::Duration => unimplemented!(),
-    };
-
-    if field.nullable {
-        if field.typ != FieldType::Binary {
-            format!("Nullable({})", typ)
-        } else {
-            warn!("Binary field cannot be nullable, ignoring nullable flag");
-            typ.to_string()
-        }
-    } else {
-        typ.to_string()
-    }
 }
